backend/requirements.txt
fastapi
uvicorn
mangum
motor
pymongo
pandas
openpyxl
google-generativeai
python-dotenv
pydantic
pydantic-settings
authlib
httpx
python-jose[cryptography]
boto3
reportlab
numpy
python-multipart
